TXMNT-597 ensure HTML response filtered for frontend
Following upgrade to Play 2.5, filtering of HTML response in implicit audit for frontend stopped working because, with Netty, the content type is inferred and headers set AFTER the filters (which actually makes sense). However, we should still be able to get the intended content type from the result entity (which is what netty is using to set Content-Type header).