[TR] TXM-3192 adds sendExplicitAudit method
Aims to simplify the creation of explicit audit events by supplying a method
which only needs the data specific to the audit event being created.

Tags are populated by the method using the implicit parameters supplied.

The README is updated to reflect the fact that new teams trying to find out how
to perform auditing will already depend on bootstrap-play-25 which simplifies
the creation of AuditConnector.