[TR] TXM-3188 Revert execution context change for release
A ticket has been created to refix: TXM-3206

The Audit class which wraps AuditConnector and the AuditingHook class
used to create implicit auditing of outbound calls both use the default
scala thread pool for calls to datastream.

This was probably never intentional as the scala thread pool is not usually
used in play microservices.

However, it has always been used in the case of the Audit class and has been
used for 1 year in the case of the AuditingHook class.

Apart from the unnecessary overhead of an additional theadpool, the use of the
scala threadpool means https://github.com/rishabh9/mdc-propagation-dispatcher can
not be used to add additonal context to all log messages (like Request-ID).

Unfortunatly the use of blocking IO in the HttpHandler class means that switching
the execution context is not necessarily safe, so this fix is being reverted
until the impact can be tested.

Also, despite the name of this project, it should not depend on play as this complicates
play upgrades as there would need to be builds of this library for the different play
versions.