[TR] TXM-3316 set user-agent
Sets the user-agent when making calls to datastream.

This is needed when diagnosing issues as kibana shows the user-agent value.

This would have been set automatically when http-verbs was used
for the call to datastream. http-verbs was removed to remove the
dependency on play-framework, and the user-agent was not explicity set.