DLK-762 Fix final counter metric
Works around a bug in the MDTP metrics which causes the metrics published at shutdown
to be averaged with any scheduled publication values.

This change insures that the correct final count is published and will never be averaged
with other values as it it not published until the value is known.