DLK-714 publish audit counters to stdout
Publishing audit counters using slf4j logging only works if the logging level is correct

In practise most microservices have their root log level set to WARN in production

Global property overrides are possible, but log levels can only be changed on a per-logger
basis if every microservice defines the logger and reads the system property

The least worst option is to write the counters directly to std out.
The message is written using similar json to the default HMRC JsonEncoder so that app, logger,
timestamp & message are parsed.
The logger is named play-auditing to help identify the source of the log message