dlk-832: remove codahale/kenshoo references
This approach brings too much fragility when bootstrap-play upgrades.
The version they're using might be binary incompatible with the version
that we're using, and it's a known difficult problem to detect.

Also, we _may_ have metrics disabled, which throws if you try to
'register' in the bootstrap-play.  It seems more elegant for
bootstrap-play to be able to provide us a no-op implementation in that
case.