dlk-832: move metricsKey into datastream metrics
In bootstrap-play, the metricsKey is not always available, and
constructing a DatastreamMetrics is not always valid.

We therefore need to extract the metricsKey and counters dependent on if
graphite is enabled or disabled.  The simpliest way I've found to do
this is inside the GraphiteMetricsModule, and we don't want to have to
instantiate the entire AuditingConfig conditionally within that module.

This split seems to make more sense.